Rework of the Texas Health 365 Fund website
The site is redone using CSS Grid so as to take advantage of positioning
and ordering of elements while still being fully committed to a
responsive and fluid design. All of the JavaScript is also reworked
using just plain, vanilla JavaScript, and removing jQuery so as to make
the page lighter. The JavaScript uses the new IntersectionObserver API
to try and off load some of the work done while scrolling the page. The
slideshow is the only piece missing from this page and may be added in
at a later date.