Update styles for Edge and no referrer for links
This commit moves some styles from HTML to the stylesheet in regards to
the footer. It also updates the footer styles to display better in Edge.

External links are now marked to have no referrer when supported and
noopener so that the new page cannot access the parent page via JavaScript when opened in a new window or tab.