Update main.css to fix Edge bug related to z-index
This fixes an Edge bug where `position: relative` needs to be added in order for the z-index on the slides to be respected. Without this, the last slide would be visible for one loop through the entire animation. Only after going through it once would the slideshow appear as intended.